work on services synchronisation

+++ ressource/ressource_client.go
@@ -4,10 +4,10 @@ import (
 	"strconv"
 
 	"io/ioutil"
-	//	"log"
 	"os"
 
 	"github.com/davecourtois/Globular/api"
+	"github.com/davecourtois/Utility"
 	"google.golang.org/grpc"
 )
 
@@ -163,10 +163,13 @@ func (self *Ressource_Client) DeleteAccount(name string) error {
 
 // Authenticate a user.
 func (self *Ressource_Client) Authenticate(name string, password string) (string, error) {
+	// In case of other domain than localhost I will rip off the token file
+	// before each authentication.
 	path := os.TempDir() + string(os.PathSeparator) + self.GetDomain() + "_token"
-
-	// remove the file if it already exist.
-	os.Remove(path)
+	if !Utility.IsLocal(self.GetDomain()) {
+		// remove the file if it already exist.
+		os.Remove(path)
+	}
 
 	rqst := &AuthenticateRqst{
 		Name:     name,
@@ -180,9 +183,11 @@ func (self *Ressource_Client) Authenticate(name string, password string) (string
 
 	// Here I will save the token into the temporary directory the token will be valid for a given time (default is 15 minutes)
 	// it's the responsability of the client to keep it refresh... see Refresh token from the server...
-	err = ioutil.WriteFile(path, []byte(rsp.Token), 0644)
-	if err != nil {
-		return "", err
+	if !Utility.IsLocal(self.GetDomain()) {
+		err = ioutil.WriteFile(path, []byte(rsp.Token), 0644)
+		if err != nil {
+			return "", err
+		}
 	}
 
 	return rsp.Token, nil
